feat: 새로 구현한 기능 도메인명 변경 적용
