+++ backend/src/main/java/com/staccato/staccato/repository/StaccatoImageRepository.java
@@ -1,10 +1,12 @@
 package com.staccato.staccato.repository;
 
 import java.util.List;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
+
 import com.staccato.staccato.domain.StaccatoImage;
 
 public interface StaccatoImageRepository extends JpaRepository<StaccatoImage, Long> {
@@ -16,4 +18,6 @@ public interface StaccatoImageRepository extends JpaRepository<StaccatoImage, Lo
     @Modifying
     @Query("DELETE FROM StaccatoImage si WHERE si.id In :ids")
     void deleteAllByIdInBulk(@Param("ids") List<Long> ids);
+
+    List<StaccatoImage> findAllByStaccatoId(@Param("staccatoId") long staccatoId);
 }
