+++ backend/src/main/resources/templates/share.html
@@ -13,9 +13,9 @@
 </head>
 <body>
 <div class="container">
-    <div class="memory-header">
-        <h1 class="memory-title"></h1>
-        <p class="memory-subtitle"></p>
+    <div class="staccato-header">
+        <h1 class="staccato-title"></h1>
+        <p class="staccato-subtitle"></p>
     </div>
 
     <div class="image-slider">
